requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.2
numpy>=1.26
loguru==0.7.2